Scan sources
The g-ir-scanner accepts many options common to the C compiler and
linker. Most notably, the library to scan can be specified with the
exact same parameter as gcc uses to pull in a dynamically linked
dependency. So the composition of the library argument, the variable
values and templates that compose is heavily inspired by the code in the
ccroot and gcc waf tools.