Allow integrating gir feature into compilation
Introspection data is often built and shipped with the library it
describes. So it is an often useful shortcut to direct both together.
Finding the link_task is actually easier when it is readily available in
the current generator.

As linker processes can be various, it is assumed that a link_task will
generate a product introspection can depend upon.